refactor(rag): Replace inlined namespace filtering logic with reusable utility
Extracted namespace filtering logic into `filter_retrievers_by_namespace` utility for consistency and reuse. Removed redundant `BucketNamespacePair` from RAGAgent. Updated event handling in ExpertRAGAgent to support `NamespaceAwareStartEvent`.
