+++ aihub_agent/aihub_agent/agents/ExpertRagAgent/ExpertRAGAgent.py
@@ -1,4 +1,5 @@
 from aihub_lib.displayers.EventDisplayer import EventDisplayer
+from aihub_lib.generative_ai.utils.filter_retrievers_by_namespace import filter_retrievers_by_namespace
 from aihub_lib.generative_ai.utils.format_expert_conversation import format_expert_conversation
 from aihub_lib.i18n.LocaleHandler import LocaleHandler
 from aihub_lib.i18n.LocaleString import LocaleString
@@ -29,6 +30,7 @@ from aihub_agent.agents.RagAgent.events.ContextInsufficientWithQueryEvent import
 from aihub_agent.agents.RagAgent.events.ExpertAnswerContextEvent import ExpertAnswerContextEvent
 from aihub_agent.agents.RagAgent.events.InOrderNodeCombinerEvent import InOrderNodeCombinerEvent
 from aihub_agent.agents.RagAgent.events.LimitChatHistoryWithContextEvent import LimitChatHistoryWithContextEvent
+from aihub_agent.agents.RagAgent.events.NamespaceAwareStartEvent import NamespaceAwareStartEvent
 from aihub_agent.agents.RagAgent.events.UserRequestsExpertEvent import UserRequestsExpertEvent
 from aihub_agent.context.run.RunContext import RunContext
 from aihub_agent.rag.preconditions import (
@@ -116,7 +118,7 @@ class ExpertRAGAgent(Agent):
     )
     async def limit_chat_history_step(
         self,
-        event: UserMessageEvent,
+        event: UserMessageEvent | NamespaceAwareStartEvent,
         agent_config: ExpertRAGAgentConfig,
     ) -> LimitChatHistoryEvent:
         return do_limit_chat_history(event.messages, agent_config.number_of_input_tokens)
@@ -128,7 +130,7 @@ class ExpertRAGAgent(Agent):
     async def condense_standalone_question_step(
         self,
         event: LimitChatHistoryEvent,
-        start_event: UserMessageEvent,
+        start_event: UserMessageEvent | NamespaceAwareStartEvent,
         agent_config: ExpertRAGAgentConfig,
         t: LocaleHandler,
         displayer: EventDisplayer,
@@ -160,11 +162,16 @@ class ExpertRAGAgent(Agent):
         self,
         event: StandaloneQuestionCondenserEvent | ContextInsufficientWithQueryEvent,
         _: FewShotAcceptEvent,
+        start_event: UserMessageEvent | NamespaceAwareStartEvent,
         agent_config: ExpertRAGAgentConfig,
         t: LocaleHandler,
     ) -> RetrieverEvent:
         """Retrieves relevant nodes from multiple knowledge sources in parallel."""
-        return await do_retrieve(event, agent_config.retrievers, t)
+        if isinstance(start_event, NamespaceAwareStartEvent):
+            retrievers = filter_retrievers_by_namespace(agent_config.retrievers, start_event.selected_namespaces)
+        else:
+            retrievers = agent_config.retrievers
+        return await do_retrieve(event, retrievers, t)
 
     @step(
         name=LocaleString(en="Rerank Retrieved Nodes"),
@@ -234,7 +241,7 @@ class ExpertRAGAgent(Agent):
         context_event: InOrderNodeCombinerEvent | ExpertAnswerContextEvent,
         chat_history_event: LimitChatHistoryEvent,
         _: ContextSufficientAcceptEvent | None,
-        start_event: UserMessageEvent,
+        start_event: UserMessageEvent | NamespaceAwareStartEvent,
         agent_config: ExpertRAGAgentConfig,
     ) -> LimitChatHistoryWithContextEvent:
         return do_limit_chat_history_with_context(
@@ -287,7 +294,7 @@ class ExpertRAGAgent(Agent):
     )
     async def forward_to_expert_asking_agent_step(
         self,
-        user_message_event: UserMessageEvent,
+        user_message_event: UserMessageEvent | NamespaceAwareStartEvent,
         _: UserRequestsExpertEvent,
         displayer: EventDisplayer,
         agent_config: ExpertRAGAgentConfig,
