+++ aihub_agent/aihub_agent/agents/RagAgent/events/__init__.py
@@ -1,4 +1,5 @@
-from aihub_agent.agents.RagAgent.events.BucketNamespacePair import BucketNamespacePair
+from aihub_lib.generative_ai.retrievers.BucketNamespacePair import BucketNamespacePair
+
 from aihub_agent.agents.RagAgent.events.ExpertAnswerContextEvent import ExpertAnswerContextEvent
 from aihub_agent.agents.RagAgent.events.NamespaceAwareStartEvent import NamespaceAwareStartEvent
